% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPlots.R
\name{dataPlots}
\alias{dataPlots}
\title{importFrom dplyr mutate select rename rename_with everything
importFrom ggplot2 ggplot aes geom_histogram element_rect stat_function stat_density scale_colour_manual scale_linetype_manual guides guide_legend element_blank
importFrom skimr skim
importFrom gridExtra grid.arrange
importFrom knitr kable
importFrom kableExtra kable_styling}
\usage{
dataPlots(df)
}
\description{
importFrom dplyr mutate select rename rename_with everything
importFrom ggplot2 ggplot aes geom_histogram element_rect stat_function stat_density scale_colour_manual scale_linetype_manual guides guide_legend element_blank
importFrom skimr skim
importFrom gridExtra grid.arrange
importFrom knitr kable
importFrom kableExtra kable_styling
}
\keyword{internal}
