% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{generate_flow_maps}
\alias{generate_flow_maps}
\title{Function for generating flow maps and animations based on transition probabilities}
\usage{
generate_flow_maps(
  hvt_model_output,
  transition_probability_df,
  hvt_plot_output,
  df,
  animation = NULL,
  flow_map = NULL,
  animation_speed = NULL,
  threshold = NULL,
  cellid_column,
  time_column
)
}
\arguments{
\item{hvt_model_output:}{Output from a Hidden Markov Model (HMM) or other modeling approach}

\item{transition_probability_df:}{Dataframe containing transition probability information}

\item{hvt_plot_output:}{Output from a Hidden Variable Transition (HVT) model}

\item{df:}{Input dataframe containing data for flow map generation}

\item{animation:}{Type of animation ('state_based', 'time_based', or 'All')}

\item{flow_map:}{Type of flow map ('with_self_state', 'without_self_state', 'self_state', 'probability', or 'All')}

\item{animation_speed:}{Speed of animation (frames per second)}

\item{threshold:}{Probability threshold for flow map arrows}

\item{cellid_column:}{Name of the column containing cell IDs}

\item{time_column:}{Name of the column containing timestamps}
}
\value{
A list of plot objects representing flow maps and animations
}
\description{
generate_flow_maps - Create Flow map
}
