% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{get_transition_probability_table}
\alias{get_transition_probability_table}
\title{Function for calculating transition probabilities for each Cell.ID in a dataframe}
\usage{
get_transition_probability_table(df, cellid_column, time_column)
}
\arguments{
\item{df:}{Input dataframe containing data for transition probabilities}

\item{cellid_column:}{Name of the column containing cell IDs}

\item{time_column:}{Name of the column containing timestamps}
}
\value{
Prints and stores a list of dataframes with transition probabilities
}
\description{
get_transition_probability_table - Create Transition Probability Table
}
