% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellId.R
\name{getCellId}
\alias{getCellId}
\title{Function to generate cell ID based on 1D sammons projection}
\usage{
getCellId(hvt.results, seed = 123)
}
\arguments{
\item{hvt.results}{List. A list of hvt.results obtained from the trainHVT function.}

\item{seed}{Numeric. Random Seed}
}
\value{
Object containing Cell.ID mappings for the given hvt.results list.
}
\description{
To generate cell id for the multivariate data, the data is being projected from n-dimensions to 1-dimension 
and the cell id is being assigned by ordering these values and finding the corresponding indexes. The output CellID gets
appended to the HVT model.
}
\details{
Cell ID
}
\examples{
data("EuStockMarkets")
hvt.results <- list()
hvt.results <- trainHVT(EuStockMarkets, n_cells = 15, depth = 1, quant.err = 0.2, 
                   distance_metric = "L1_Norm", error_metric = "mean",
                   projection.scale = 10, normalize = TRUE, seed = 123,
                   quant_method="kmeans",diagnose=TRUE)
plotHVT(hvt.results, line.width = c(0.8), color.vec = c('#141B41'), 
        maxDepth = 1, heatmap ='1D')
getCellId (hvt.results)
}
\author{
Shubhra Prakash <shubhra.prakash@mu-sigma.com>
}
