% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{muHVT}
\alias{muHVT}
\alias{muHVT-package}
\alias{muHVT-package}
\title{muHVT - Hierarchical Voronoi Tessellations}
\description{
Constructing hierarchical voronoi tessellations for a given data set and
overlay heatmaps for variables at various levels of the tessellations for
in-depth data analysis.
}
\details{
\tabular{ll}{ Package: \tab muHVT\cr Type: \tab Package\cr Version: \tab
1.0.0\cr Date: \tab 2020-01-06\cr License: \tab GPL-2 | GPL-3\cr }
}
\seealso{
\code{\link{HVT}}
}
\author{
Sangeet Moy Das, Zubin Dowlaty, Avinash Joshi, Meet Dave

Maintainer: Mu Sigma Inc. <ird.experiencelab@mu-sigma.com>
}
\keyword{package}
