% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCentroids.R, R/ggplotTessHmap.R,
%   R/ggplotTileHmap.R, R/predictHVT.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%>\%}
\alias{ggplotTessHmap}
\alias{\%>\%}
\alias{ggplotTileHmap}
\alias{\%>\%}
\alias{predictHVT}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

lhs \%>\% rhs

lhs \%>\% rhs

lhs \%>\% rhs
}
\description{
Pipe operator

Pipe operator

Pipe operator

Pipe operator
}
\keyword{internal}
