% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiag.R
\name{plotDiag}
\alias{plotDiag}
\title{Make the diagnostic plots for hierarchical voronoi tessellations model.}
\usage{
plotDiag(model_obj)
}
\arguments{
\item{model_obj}{List. A list of model_obj obtained from the HVT
function or prediction object}
}
\description{
Make the diagnostic plots for hierarchical voronoi tessellations model.
}
\seealso{
\code{\link{plotHVT}}
}
\author{
Shubhra Prakash <shubhra.prakash@mu-sigma.com>
}
\keyword{diagnostics}
