% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHVT.R
\name{plotHVT}
\alias{plotHVT}
\title{Plot the hierarchical tessellations.}
\usage{
plotHVT(
  hvt.results,
  line.width,
  color.vec,
  pch1 = 21,
  palette.color = 6,
  centroid.size = 1.5,
  title = NULL,
  maxDepth = NULL,
  dataset,
  child.level,
  hmap.cols,
  previous_level_heatmap = TRUE,
  show.points = FALSE,
  asp = 1,
  ask = TRUE,
  tess.label = NULL,
  quant.error.hmap = NULL,
  n_cells.hmap = NULL,
  label.size = 0.5,
  sepration_width = 7,
  layer_opacity = c(0.5, 0.75, 0.99),
  dim_size = 1000,
  heatmap = "2DHVT"
)
}
\arguments{
\item{hvt.results}{(1D/2D/3D) List. A list containing the ouput of \code{trainHVT} function
which has the details of the tessellations to be plotted.}

\item{line.width}{(2D) Numeric Vector. A vector indicating the line widths of the
tessellation boundaries for each level.}

\item{color.vec}{(2D) Vector. A vector indicating the colors of the boundaries of
the tessellations at each level.}

\item{pch1}{(2D) Numeric. Symbol type of the centroids of the tessellations
(parent levels). Default value is 21.}

\item{palette.color}{(2DHEATMAP) Numeric. A number indicating the heat map color palette. 1 - rainbow,
2 - heat.colors, 3 - terrain.colors, 4 - topo.colors, 5 - cm.colors, 
6 - BlCyGrYlRd (Blue,Cyan,Green,Yellow,Red) color (default = 6).}

\item{centroid.size}{(2D) Numeric. Size of centroids of first level
tessellations.}

\item{title}{(2DHVT) String. Set a title for the plot. (default = NULL)}

\item{maxDepth}{(2DHVT) Numeric. An integer indicating the number of levels. (default = NULL)}

\item{dataset}{(2DHEATMAP) Data frame. The input data set.}

\item{child.level}{(2DHEATMAP/3D) Numeric. Indicating the level for which the heat map is
to be plotted.}

\item{hmap.cols}{(2DHEATMAP/3D) Numeric or Character. The column number of column name from
the dataset indicating the variables for which the heat map is to be
plotted.}

\item{previous_level_heatmap}{(2DHEATMAP) Logical. If TRUE, the heatmap of previous level
will be overlayed on the heatmap of selected level. If #' FALSE, the heatmap
of only selected level will be plotted}

\item{show.points}{(2DHEATMAP) Logical. Indicating if the centroids should
be plotted on the tessellations. (default = FALSE)}

\item{asp}{(2DHEATMAP) Numeric. Indicating the aspect ratio type. For flexible aspect
ratio set, asp = NA. (default = 1)}

\item{ask}{(2DHEATMAP) Logical. If TRUE (and the R session is interactive) the user is
asked for input, before a new figure is drawn. (default = TRUE)}

\item{tess.label}{(2DHEATMAP) Vector. A vector for labelling the tessellations. (default
= NULL)}

\item{quant.error.hmap}{(2DHEATMAP) Numeric. A number indicating the quantization error
threshold.}

\item{n_cells.hmap}{(2DHEATMAP/3D) Numeric. An integer indicating the number of
cells/clusters per hierarchy (level)}

\item{label.size}{(2DHEATMAP) Numeric. The size by which the tessellation labels should
be scaled. (default = 0.5)}

\item{sepration_width}{(3D) Numeric. An integer indicating the width between two Levels}

\item{layer_opacity}{(3D) Numeric. A vector indicating the opacity of each layer/ level}

\item{dim_size}{(3D) Numeric. An integer indicating the dimension size used to create the matrix for the plot}

\item{heatmap}{Character. An option to indicate which type of plot should be generated. Accepted entries are 
'1D', '2DHVT','2DHEATMAP', '3D'. Default value is 2DHVT.}
}
\value{
plot object containing the main HVT plot for the given HVT results and heatmap type.
}
\description{
This is the main plotting function to construct hierarchical voronoi tessellations in 1D or 2D or 3D.
}
\examples{
dataset <- data.frame(date = as.numeric(time(EuStockMarkets)),
                     DAX = EuStockMarkets[, "DAX"],
                     SMI = EuStockMarkets[, "SMI"],
                     CAC = EuStockMarkets[, "CAC"],
                     FTSE = EuStockMarkets[, "FTSE"])
dataset_hvt <- dataset[,-c(1)]
hvt.results <- list()
hvt.results <- trainHVT(dataset_hvt, n_cells = 15, depth = 1, quant.err = 0.2, 
                       distance_metric = "L1_Norm", error_metric = "mean",
                       projection.scale = 10, normalize = TRUE, seed = 123,
                       quant_method="kmeans")
                       
#1D - Plot                        
plotHVT(hvt.results, heatmap='1D')                       
#2D - HVT Plot
plotHVT(hvt.results, line.width = c(1.2), color.vec = c('#000000'), pch = 21, centroid.size = 1, 
maxDepth = 1,heatmap = '2DHVT')
#2D - HEATMAP
plotHVT(hvt.results, EuStockMarkets, centroid.size = 1,
child.level = 1, hmap.cols = "DAX",
line.width = c(0.6), color.vec = ('#000000') , 
pch1 = 21, heatmap = '2DHEATMAP')
#3D - Plot
plotHVT( hvt.results, child.level = 1, hmap.cols = "DAX", n_cells.hmap = 15, 
layer_opacity = c(0.7, 0.8, 0.99), dim_size = 1000, heatmap = '3D' )
}
\seealso{
\code{\link{trainHVT}}
}
\author{
Shubhra Prakash <shubhra.prakash@mu-sigma.com>, Sangeet Moy Das <sangeet.das@mu-sigma.com>
}
\keyword{Tessellation_and_Heatmap}
