% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantErrorHistogram.R
\name{plotQuantErrorHistogram}
\alias{plotQuantErrorHistogram}
\title{Make the diagnostic plots for hierarchical voronoi tessellations model.}
\usage{
plotQuantErrorHistogram(hvt.results, hvt.scoring)
}
\arguments{
\item{hvt.results}{List. A list of hvt.results obtained from the trainHVT
function.}

\item{hvt.scoring}{List. A list of hvt.scoring obtained from the scoreHVT
function.}
}
\value{
Returns the ggplot object containing the Quantized Error distribution plots for the given HVT results and scores.
}
\description{
This is the function that produces histograms displaying the distribution of Quantized Error (QE) values 
for both train and test datasets, highlighting mean values with dashed lines for quick evaluation.
}
\examples{
data("EuStockMarkets")
dataset <- data.frame(date = as.numeric(time(EuStockMarkets)),
                     DAX = EuStockMarkets[, "DAX"],
                     SMI = EuStockMarkets[, "SMI"],
                     CAC = EuStockMarkets[, "CAC"],
                     FTSE = EuStockMarkets[, "FTSE"])
#adding this step especially for this function
rownames(EuStockMarkets) <- dataset$date
#Split in train and test
train <- EuStockMarkets[1:1302, ]
test <- EuStockMarkets[1303:1860, ]
hvt_summary<- trainHVT(train,n_cells = 15, depth = 1, quant.err = 0.2,
                      distance_metric = "L1_Norm", error_metric = "mean",
                      projection.scale = 10, normalize = TRUE,seed = 123,
                      quant_method = "kmeans")
scoring <- scoreHVT(test, hvt_summary, child.level = 2, mad.threshold = 0.2)
plotQuantErrorHistogram(hvt_summary, scoring)  
}
\seealso{
\code{\link{plotHVT}}
}
\author{
Shubhra Prakash <shubhra.prakash@mu-sigma.com>
}
\keyword{Diagnostics_or_Validation}
