% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStateTransitions.R
\name{plotStateTransitions}
\alias{plotStateTransitions}
\title{Creating State Transition Plot}
\usage{
plotStateTransitions(
  df,
  sample_size = NULL,
  line_plot = NULL,
  cellid_column,
  time_column
)
}
\arguments{
\item{df}{Data frame. Input dataframe should contain two columns of cell ID from scoreHVT function and timestamp.}

\item{sample_size}{Numeric. An integer indicating the Fraction of the dataframe to sample.
Default value is 0.2.}

\item{line_plot}{Logical. A logical value indicating to create a line plot.
Default value is NULL.}

\item{cellid_column}{character. Name of the column containing cell IDs.}

\item{time_column}{character. Name of the column containing timestamps.}
}
\value{
A plotly object representing the state transition plot for the given dataframe.
}
\description{
Main function to creating state transition plot from a dataframe.
}
\details{
This is the main function to creating state transition plot from a dataframe.
A state transition plot is a type of data visualization used to represent
the changes or transitions in states over time for a given system.
State refers to a particular condition or status of a cell at a specific point in time.
This could be related to various attributes like a phase in a cycle, or any other measurable characteristic.
Transition refers to the change of state for a cell from one condition to another over time.
This could represent processes like progression through stages, changes in activity, or other dynamic characteristics.
}
\examples{
dataset <- data.frame(
  date = as.numeric(time(EuStockMarkets)),
  DAX = EuStockMarkets[, "DAX"],
  SMI = EuStockMarkets[, "SMI"],
  CAC = EuStockMarkets[, "CAC"],
  FTSE = EuStockMarkets[, "FTSE"]
)
dataset_hvt <- dataset[, -c(1)]
hvt.results <- list()
hvt.results <- trainHVT(dataset_hvt,
  n_cells = 15, depth = 3, quant.err = 0.2,
  distance_metric = "L1_Norm", error_metric = "mean",
  projection.scale = 10, normalize = TRUE, seed = 123,
  quant_method = "kmeans"
)

predictions <- scoreHVT(dataset_hvt, hvt.results, child.level = 2, mad.threshold = 0.2)
cell_id <- predictions$scoredPredictedData$Cell.ID
time_stamp <- dataset$date
dataset <- data.frame(cell_id, time_stamp)
plotStateTransitions(dataset,
  sample_size = NULL,
  line_plot = FALSE,
  cellid_column = "cell_id",
  time_column = "time_stamp"
)
}
\seealso{
\code{\link{trainHVT}} \cr \code{\link{scoreHVT}}
}
\author{
PonAnuReka Seenivasan <ponanureka.s@mu-sigma.com>
}
\keyword{/}
\keyword{Outliers}
\keyword{Transition}
