% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qeHistPlot.R
\name{qeHistPlot}
\alias{qeHistPlot}
\title{plotDiag

Make the diagnostic plots for hierarchical voronoi tessellations model.}
\usage{
qeHistPlot(hvt.results, hvt.predictions)
}
\arguments{
\item{hvt.results}{List. A list of hvt.results obtained from the HVT
function.}

\item{hvt.predictions}{List. A list of hvt.predictions obtained from the Predict
function.}
}
\description{
plotDiag

Make the diagnostic plots for hierarchical voronoi tessellations model.
}
\seealso{
\code{\link{plotHVT}}
}
\author{
Shubhra Prakash <shubhra.prakash@mu-sigma.com>
}
\keyword{diagnostics}
