% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{reconcile_transition_probability}
\alias{reconcile_transition_probability}
\title{Function for creating transition probability heatmaps and reconciling them}
\usage{
reconcile_transition_probability(
  df,
  hmap_type = NULL,
  cellid_column,
  time_column
)
}
\arguments{
\item{df:}{Input dataframe containing data for transition probabilities}

\item{hmap_type:}{Type of heatmap to generate ('with_self_state', 'without_self_state', or 'All')}

\item{cellid_column:}{Name of the column containing cell IDs}

\item{time_column:}{Name of the column containing timestamps}
}
\value{
A list of plotly heatmap objects representing the transition probability heatmaps
}
\description{
reconcile_transition_probability - Function for Reconciliation of Transition Probability - Model Diagnostics
}
