% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{sampledata - Sample dataset to run Jensen-Shannon-Bregman distance}
\format{A data frame with 4000 observations on the following 2 variables.
\describe{ \item{GLD}{ a numeric vector } \item{USO}{a numeric vector } }}
\usage{
sampledata
}
\description{
A dataset consisting of two timeseries equities
}
\details{
A dataset containing two timeseries equities : GLD and USO
}
\examples{

data(sampledata)

}
\keyword{datasets}
