% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowmap.R
\name{state_transition_plot}
\alias{state_transition_plot}
\title{Function for creating state transition plots from a dataframe}
\usage{
state_transition_plot(
  df,
  sample_size = NULL,
  line_plot = NULL,
  cellid_column,
  time_column
)
}
\arguments{
\item{df:}{Input dataframe containing data for state transitions}

\item{sample_size:}{Fraction of the dataset to sample (default is 0.2)}

\item{line_plot:}{Boolean flag to determine whether to create a line plot (default is NULL)}

\item{cellid_column:}{Name of the column containing cell IDs}

\item{time_column:}{Name of the column containing timestamps}
}
\value{
A plotly plot object representing the state transition plot
}
\description{
state_transition_plot - Create Timeseries Plot
}
