% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edaPlots.R
\name{edaPlots}
\alias{edaPlots}
\title{plots for understanding dataset}
\usage{
edaPlots(df, time_series = FALSE, time_column)
}
\arguments{
\item{df}{Dataframe. The input dataset, can be a time series too.}

\item{time_series}{Logical. A value to indicate whether the dataset is time_series or not.}

\item{time_column}{Character. The name of the time column in the dataset.}
}
\value{
Five objects which includes time series plots, data distribution plots, 
box plots, correlation matrix plot and a descriptive statistics table.
}
\description{
main function that gives all the eda plots
}
\examples{
dataset <- data.frame(date = as.numeric(time(EuStockMarkets)),
DAX = EuStockMarkets[, "DAX"],
SMI = EuStockMarkets[, "SMI"],
CAC = EuStockMarkets[, "CAC"],
FTSE = EuStockMarkets[, "FTSE"])
edaPlots(dataset, time_series = TRUE, time_column = 'date')
}
\author{
Vishwavani <vishwavani@mu-sigma.com>
}
\keyword{EDA}
