% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellId.R
\name{getCellId}
\alias{getCellId}
\title{Cell ID}
\usage{
getCellId(hvt.results, seed = 123)
}
\arguments{
\item{hvt.results}{List. A list of hvt.results obtained from the trainHVT function.}

\item{seed}{Numeric. Random Seed}
}
\value{
Object containing Cell.ID mappings for the given hvt.results list.
}
\description{
Function to generate cell ID based on 1D sammons projection
}
\details{
To generate cell id for the multivariate data, the data is being projected from n-dimensions to 1-dimension 
and the cell id is being assigned by ordering these values and finding the corresponding indexes. The output Cell id gets
appended to the HVT model.
}
\author{
Shubhra Prakash <shubhra.prakash@mu-sigma.com>
}
\keyword{internal}
