% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransitionProbability.R
\name{getTransitionProbability}
\alias{getTransitionProbability}
\title{Creating Transition Probability table}
\usage{
getTransitionProbability(
  df,
  cellid_column,
  time_column,
  all_prob_results = NULL
)
}
\arguments{
\item{df}{Data frame. Input dataframe should contain two columns of cell ID from scoreHVT function and timestamp.}

\item{cellid_column}{Character. Name of the column containing cell IDs.}

\item{time_column}{Character. Name of the column containing timestamps.}

\item{all_prob_results}{Logical. A Logical value indicating the facility to print the results. TRUE in R script saves the list of dataframe as environment variable and print the results only.
FALSE in rmd file by giving chunk option as {result = 'asis'} prints the probability table as dropdown.}
}
\value{
Prints and stores a list of dataframes with transition probabilities.
}
\description{
This is the main function to create transition probability table.
The transition probability table quantifies the likelihood of transitioning from one state to another. 
It is derived from historical data and provides a statistical basis for predicting future state changes.
States: The table includes the current states and the possible next states.
Probabilities: For each current state, it lists the probability of transitioning to each of the next possible states. 
Frequency: It often includes the frequency of these transitions, which helps in understanding the commonness of certain state changes.
}
\examples{
dataset <- data.frame(date = as.numeric(time(EuStockMarkets)),
DAX = EuStockMarkets[, "DAX"],
SMI = EuStockMarkets[, "SMI"],
CAC = EuStockMarkets[, "CAC"],
FTSE = EuStockMarkets[, "FTSE"])
dataset_hvt <- dataset[,-c(1)]
hvt.results <- list()
hvt.results <- trainHVT(dataset_hvt, n_cells = 15, depth = 3, quant.err = 0.2, 
                   distance_metric = "L1_Norm", error_metric = "mean",
                   projection.scale = 10, normalize = TRUE, seed = 123,
                   quant_method="kmeans")
predictions <- scoreHVT(dataset_hvt, hvt.results, child.level = 2, mad.threshold = 0.2) 
cell_id <- predictions$scoredPredictedData$Cell.ID
time_stamp <- dataset$date
dataset <- data.frame(cell_id, time_stamp)
getTransitionProbability(dataset,cellid_column = "cell_id", time_column = "time_stamp")
}
\seealso{
\code{\link{trainHVT}} \cr \code{\link{scoreHVT}}
}
\author{
PonAnuReka Seenivasan <ponanureka.s@mu-sigma.com>
}
\keyword{internal}
