% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muHVT-package.R
\docType{package}
\name{muHVT}
\alias{muHVT}
\alias{muHVT-package}
\alias{muHVT-package}
\title{muHVT - Hierarchical Voronoi Tessellations}
\description{
Constructing hierarchical voronoi tessellations for a given data set and
overlay heatmaps for variables at various levels of the tessellations for
in-depth data analysis.
}
\details{
\tabular{ll}{ Package: \tab muHVT\cr Type: \tab Package\cr Version: \tab
1.0.0\cr Date: \tab 2013-07-10\cr License: \tab LGPL version 2\cr }
}
\seealso{
\code{\link{HVT}}
}
\author{
Vignesh C Prabhu, Pravin V, Karthik Devaraj

Maintainer: Karthik Devaraj <karthik.devaraj@mu-sigma.com>
}
\keyword{package}
