% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnimatedFlowmap.R
\name{plotAnimatedFlowmap}
\alias{plotAnimatedFlowmap}
\title{Generating flow maps and animations based on transition probabilities}
\usage{
plotAnimatedFlowmap(
  hvt_model_output,
  transition_probability_df,
  hvt_plot_output,
  df,
  animation = NULL,
  flow_map = NULL,
  animation_speed = NULL,
  threshold = NULL,
  cellid_column,
  time_column
)
}
\arguments{
\item{hvt_model_output}{List. Output from a trainHVT function.}

\item{transition_probability_df}{Data frame. Output Dataframe from getTransitionProbability function}

\item{hvt_plot_output}{List. Base plot for the flow maps.}

\item{df}{Data frame. Input dataframe should contain two columns of cell ID from scoreHVT function and timestamp.}

\item{animation}{character. Type of animation ('state_based', 'time_based', or 'All').}

\item{flow_map}{character. Type of flow map ('with_self_state', 'without_self_state', 'self_state', 'probability', or 'All').}

\item{animation_speed}{Numeric. An Integer representing the Speed of animation (frames per second).}

\item{threshold}{Numeric. An Integer representing the Probability threshold for flow map arrows.}

\item{cellid_column}{character. Name of the column containing cell IDs.}

\item{time_column}{character. Name of the column containing timestamps}
}
\value{
A list of plot objects representing flow maps and animations.
}
\description{
Main function for generating flow maps and animations based on transition probabilities.
}
\details{
Flow maps are a type of data visualization used to represent movements or transitions between different locations or states. 
They visually connect points to show the direction and volume of movements, such as the transitions in a Hidden Markov Model.
These maps help in understanding the dynamics of the system being studied by visually representing the direction and magnitude of flows or transitions.
}
\examples{
dataset <- data.frame(date = as.numeric(time(EuStockMarkets)),
DAX = EuStockMarkets[, "DAX"],
SMI = EuStockMarkets[, "SMI"],
CAC = EuStockMarkets[, "CAC"],
FTSE = EuStockMarkets[, "FTSE"])
dataset_hvt <- dataset[,-c(1)]
hvt.results <- list()
hvt.results <- trainHVT(dataset_hvt, n_cells = 15, depth = 1, quant.err = 0.2, 
                   distance_metric = "L1_Norm", error_metric = "mean",
                   projection.scale = 10, normalize = TRUE, seed = 123,
                   quant_method="kmeans")
predictions <- scoreHVT(dataset_hvt, hvt.results, child.level = 2, mad.threshold = 0.2) 
cell_id <- predictions$scoredPredictedData$Cell.ID
time_stamp <- dataset$date
dataset <- data.frame(cell_id, time_stamp)
getTransitionProbability(dataset,cellid_column = "cell_id", time_column = "time_stamp")
plot <- plotAnimatedFlowmap(hvt_model_output = hvt.results, transition_probability_df = trans_prob_df,  df = dataset, animation = NULL, flow_map = NULL, animation_speed = NULL, threshold = NULL,  cellid_column = "cell_id", time_column = "time_stamp") 
gif <- magick::image_read(plot[[2]])
print(plot[[1]])
print(gif)
}
\seealso{
\code{\link{trainHVT}} \cr \code{\link{scoreHVT}} \cr \code{\link{getTransitionProbability}}
}
\author{
PonAnuReka Seenivasan <ponanureka.s@mu-sigma.com>
}
\keyword{/}
\keyword{Outliers}
\keyword{Transition}
