% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantErrorHistogram.R
\name{plotQuantErrorHistogram}
\alias{plotQuantErrorHistogram}
\title{Make the diagnostic plots for hierarchical voronoi tessellations model.}
\arguments{
\item{hvt.results}{List. A list of hvt.results obtained from the trainHVT
function.}

\item{hvt.predictions}{List. A list of hvt.predictions obtained from the scoreHVT
function.}
}
\value{
Returns the ggplot object containing the Quantized Error distribution plots for the given HVT results and predictions
}
\description{
Make the diagnostic plots for hierarchical voronoi tessellations model.
}
\examples{
dataset <- data.frame(date = as.numeric(time(EuStockMarkets)),
DAX = EuStockMarkets[, "DAX"],
SMI = EuStockMarkets[, "SMI"],
CAC = EuStockMarkets[, "CAC"],
FTSE = EuStockMarkets[, "FTSE"])
dataset_hvt <- dataset[,-c(1)]
hvt_results <- list()
hvt_results <- trainHVT(dataset_hvt, n_cells = 15, depth = 1, quant.err = 0.2,
                        distance_metric = "L1_Norm", error_metric = "mean",
                        projection.scale = 10, normalize = TRUE, seed = 123,
                        quant_method="kmeans", hvt_validation = TRUE, diagnose = TRUE)
plotQuantErrorHistogram(hvt.results, predictions)  
}
\seealso{
\code{\link{plotHVT}}
}
\author{
Shubhra Prakash <shubhra.prakash@mu-sigma.com>
}
\keyword{/}
\keyword{Diagnostics}
\keyword{Validation}
