% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQuantErrorHistogram.R
\name{plotQuantErrorHistogram}
\alias{plotQuantErrorHistogram}
\title{Make the diagnostic plots for hierarchical voronoi tessellations model.}
\arguments{
\item{hvt.results}{List. A list of hvt.results obtained from the trainHVT
function.}

\item{hvt.predictions}{List. A list of hvt.predictions obtained from the scoreHVT
function.}
}
\value{
Returns the ggplot object containing the Quantized Error distribution plots for the given HVT results and predictions
}
\description{
Make the diagnostic plots for hierarchical voronoi tessellations model.
}
\seealso{
\code{\link{plotHVT}}
}
\author{
Shubhra Prakash <shubhra.prakash@mu-sigma.com>
}
\keyword{/}
\keyword{Diagnostics}
\keyword{Validation}
