% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStateTransition.R
\name{plotStateTransition}
\alias{plotStateTransition}
\title{Creating State Transition Plot}
\usage{
plotStateTransition(
  df,
  sample_size = NULL,
  line_plot = NULL,
  cellid_column,
  time_column
)
}
\arguments{
\item{df}{Data frame. Input dataframe should contain two columns. Cell ID from scoreHVT function and timestamp of that dataset.}

\item{sample_size}{Numeric. An integer indicating the Fraction of the dataframe to visualize in the plot.
Default value is 0.2}

\item{line_plot}{Logical. A logical value indicating to create a line plot. Default value is NULL.}

\item{cellid_column}{Character. Name of the column containing cell IDs.}

\item{time_column}{Character. Name of the column containing timestamps.}
}
\value{
A plotly object representing the state transition plot for the given dataframe.
}
\description{
This is the main function to creating state transition plot from a dataframe.
A state transition plot is a type of data visualization used to represent 
the changes or transitions in states over time for a given system. 
State refers to a particular condition or status of a cell at a specific point in time. 
Transition refers to the change of state for a cell from one condition to another over time.
}
\examples{
dataset <- data.frame(date = as.numeric(time(EuStockMarkets)),
DAX = EuStockMarkets[, "DAX"],
SMI = EuStockMarkets[, "SMI"],
CAC = EuStockMarkets[, "CAC"],
FTSE = EuStockMarkets[, "FTSE"])
rownames(EuStockMarkets) <- dataset$date
hvt.results<- trainHVT(dataset,n_cells = 60, depth = 1, quant.err = 0.1,
                       distance_metric = "L1_Norm", error_metric = "max",
                       normalize = TRUE,quant_method = "kmeans")
scoring <- scoreHVT(dataset, hvt.results)
cell_id <- scoring$scoredPredictedData$Cell.ID
time_stamp <- dataset$date
dataset <- data.frame(cell_id, time_stamp)
plotStateTransition(dataset, sample_size = 1, cellid_column = "cell_id",time_column = "time_stamp")
}
\seealso{
\code{\link{trainHVT}} \cr \code{\link{scoreHVT}}
}
\author{
PonAnuReka Seenivasan <ponanureka.s@mu-sigma.com>
}
\keyword{Transition_or_Prediction}
