% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcileTransitionProbability.R
\name{reconcileTransitionProbability}
\alias{reconcileTransitionProbability}
\title{Reconciliation of Transition Probability - Model Diagnostics

Main function for creating transition probability heatmaps and reconcilation of the same.

This is the main function to creating transition probability heatmaps and reconcilation of the same.
The Reconciliation of Transition Probability refers to the process of analyzing and adjusting transition probabilities in a stochastic model like a Markov Chain. 
It involves ensuring the probabilities accurately reflect real-world dynamics by normalizing them,
removing unlikely transitions, and comparing different models. 
The function described creates heatmaps to visually represent these probabilities,
aiding in the understanding and analysis of state transitions within the model.}
\arguments{
\item{df}{Data frame. Input dataframe should contain two columns of cell ID from scoreHVT function and timestamp.}

\item{cellid_column}{character. Name of the column containing cell IDs.}

\item{time_column}{character. Name of the column containing timestamps}

\item{hmap_type}{character. Type of heatmap to generate ('with_self_state', 'without_self_state', or 'All')}
}
\value{
A list of plotly heatmap objects representing the transition probability heatmaps.
}
\description{
Reconciliation of Transition Probability - Model Diagnostics

Main function for creating transition probability heatmaps and reconcilation of the same.

This is the main function to creating transition probability heatmaps and reconcilation of the same.
The Reconciliation of Transition Probability refers to the process of analyzing and adjusting transition probabilities in a stochastic model like a Markov Chain. 
It involves ensuring the probabilities accurately reflect real-world dynamics by normalizing them,
removing unlikely transitions, and comparing different models. 
The function described creates heatmaps to visually represent these probabilities,
aiding in the understanding and analysis of state transitions within the model.
}
\examples{
dataset <- data.frame(date = as.numeric(time(EuStockMarkets)),
DAX = EuStockMarkets[, "DAX"],
SMI = EuStockMarkets[, "SMI"],
CAC = EuStockMarkets[, "CAC"],
FTSE = EuStockMarkets[, "FTSE"])
dataset_hvt <- dataset[,-c(1)]
hvt.results <- list()
hvt.results <- trainHVT(dataset_hvt, n_cells = 15, depth = 3, quant.err = 0.2, 
                   distance_metric = "L1_Norm", error_metric = "mean",
                   projection.scale = 10, normalize = TRUE, seed = 123,
                   quant_method="kmeans")
predictions <- scoreHVT(dataset_hvt, hvt.results, child.level = 2, mad.threshold = 0.2) 
cell_id <- predictions$scoredPredictedData$Cell.ID
time_stamp <- dataset$date
dataset <- data.frame(cell_id, time_stamp)
reconcileTransitionProbability(dataset, hmap_type = 'All', 
cellid_column = "cell_id", time_column = "time_stamp") 
}
\seealso{
\code{\link{trainHVT}} \cr \code{\link{scoreHVT}}
}
\author{
PonAnuReka Seenivasan <ponanureka.s@mu-sigma.com>
}
\keyword{/}
\keyword{Diagnostics}
\keyword{Validation}
